requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-ollama 
faiss-cpu 
sentence-transformers
beautifulsoup4
flask
PyPDF2>=3.0.0
python-docx>=0.8.11

